+++ enjoy/src/renderer/components/preferences/hotkeys.tsx
@@ -1,17 +1,10 @@
 import { t } from "i18next";
 import {
   Separator,
-  AlertDialog,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  Button,
-  toast,
 } from "@renderer/components/ui";
-import { HotKeysSettingsProviderContext, Hotkey } from "@/renderer/context";
-import { useContext, useState, useMemo, useEffect } from "react";
+import { HotKeysSettingsProviderContext, Hotkey } from "@renderer/context";
+import { HotkeysSettings } from "@renderer/components";
+import { useContext, useState } from "react";
 
 export const Hotkeys = () => {
   const [open, setOpen] = useState(false);
@@ -36,8 +29,8 @@ export const Hotkeys = () => {
 
         <div className="flex items-center justify-between py-4">
           <div className="flex items-center space-x-2">{t("quitApp")}</div>
-          <kbd className="bg-muted px-2 py-1 rounded-md text-sm text-muted-foreground cursor-not-allowed">
-            {commandOrCtrl} + Q
+          <kbd className="bg-muted px-2 py-1 rounded-md text-sm text-muted-foreground cursor-not-allowed capitalize">
+            {commandOrCtrl}+Q
           </kbd>
         </div>
 
@@ -54,7 +47,7 @@ export const Hotkeys = () => {
                 keyName: "OpenPreferences",
               })
             }
-            className="bg-muted px-2 py-1 rounded-md text-sm text-muted-foreground cursor-pointer"
+            className="bg-muted px-2 py-1 rounded-md text-sm text-muted-foreground cursor-pointer capitalize"
           >
             {currentHotkeys.OpenPreferences}
           </kbd>
@@ -74,7 +67,7 @@ export const Hotkeys = () => {
                 keyName: "PlayOrPause",
               })
             }
-            className="bg-muted px-2 py-1 rounded-md text-sm text-muted-foreground cursor-pointer"
+            className="bg-muted px-2 py-1 rounded-md text-sm text-muted-foreground cursor-pointer capitalize"
           >
             {currentHotkeys.PlayOrPause}
           </kbd>
@@ -93,7 +86,7 @@ export const Hotkeys = () => {
                 keyName: "StartOrStopRecording",
               })
             }
-            className="bg-muted px-2 py-1 rounded-md text-sm text-muted-foreground cursor-pointer"
+            className="bg-muted px-2 py-1 rounded-md text-sm text-muted-foreground cursor-pointer capitalize"
           >
             {currentHotkeys.StartOrStopRecording}
           </kbd>
@@ -112,7 +105,7 @@ export const Hotkeys = () => {
                 keyName: "PlayOrPauseRecording",
               })
             }
-            className="bg-muted px-2 py-1 rounded-md text-sm text-muted-foreground cursor-pointer"
+            className="bg-muted px-2 py-1 rounded-md text-sm text-muted-foreground cursor-pointer capitalize"
           >
             {currentHotkeys.PlayOrPauseRecording}
           </kbd>
@@ -150,7 +143,7 @@ export const Hotkeys = () => {
                 keyName: "PlayNextSegment",
               })
             }
-            className="bg-muted px-2 py-1 rounded-md text-sm text-muted-foreground cursor-pointer"
+            className="bg-muted px-2 py-1 rounded-md text-sm text-muted-foreground cursor-pointer capitalize"
           >
             {currentHotkeys.PlayNextSegment}
           </kbd>
@@ -169,7 +162,7 @@ export const Hotkeys = () => {
                 keyName: "Compare",
               })
             }
-            className="bg-muted px-2 py-1 rounded-md text-sm text-muted-foreground cursor-pointer"
+            className="bg-muted px-2 py-1 rounded-md text-sm text-muted-foreground cursor-pointer capitalize"
           >
             {currentHotkeys.Compare}
           </kbd>
@@ -178,129 +171,12 @@ export const Hotkeys = () => {
         <Separator />
       </div>
 
-      <ChangeHotkeyDialog
+      <HotkeysSettings
         open={open}
         keyName={selectedItem?.keyName}
         name={selectedItem?.name}
         onOpenChange={setOpen}
       />
     </>
-  );
-};
-
-const ChangeHotkeyDialog = ({
-  open,
-  name,
-  keyName,
-  onOpenChange,
-}: {
-  open: boolean;
-  name: string;
-  keyName: string;
-  onOpenChange: (open: boolean) => void;
-}) => {
-  const {
-    changeHotkey,
-    currentHotkeys,
-    recordingHotkeys,
-    resetRecordingHotkeys,
-    startRecordingHotkeys,
-    stopRecordingHotkeys,
-  } = useContext(HotKeysSettingsProviderContext);
-  const [isRecording, setIsRecording] = useState(false);
-
-  const joinedKeys = useMemo(
-    () => [...recordingHotkeys].join("+"),
-    [recordingHotkeys]
-  );
-
-  const changeKeyMap = async () => {
-    const ret = (await changeHotkey(keyName, recordingHotkeys)) as unknown as {
-      error: "conflict" | "invalid";
-      data: string | string[];
-      input: string;
-    };
-    setIsRecording(false);
-    const { error, data, input } = ret ?? {};
-
-    if (error === "conflict") {
-      toast.error(
-        t("customizeShortcutsConflictToast", {
-          input,
-          otherHotkeyName: (data as string[]).join(","),
-        })
-      );
-    } else if (error === "invalid") {
-      toast.error(t("customizeShortcutsInvalidToast"));
-    } else {
-      toast.success(t("customizeShortcutsUpdated"));
-      onOpenChange(false);
-    }
-  };
-
-  const clear = () => {
-    resetRecordingHotkeys();
-    setIsRecording(false);
-  };
-
-  useEffect(() => {
-    if (isRecording) {
-      startRecordingHotkeys();
-    } else {
-      stopRecordingHotkeys();
-    }
-  }, [isRecording]);
-
-  useEffect(() => {
-    return () => {
-      setIsRecording(false);
-    };
-  }, [open]);
-
-  return (
-    <AlertDialog open={open} onOpenChange={onOpenChange}>
-      <AlertDialogContent>
-        <AlertDialogHeader>
-          <AlertDialogTitle>{name}</AlertDialogTitle>
-        </AlertDialogHeader>
-        <div>
-          {isRecording ? (
-            <div className="">
-              <div className="flex justify-center mb-4">
-                <Button variant="secondary">
-                  {joinedKeys.length > 0 ? (
-                    <span className="text-sm">{joinedKeys}</span>
-                  ) : (
-                    <span className="font-mono">-</span>
-                  )}
-                </Button>
-              </div>
-              <div className="py-2 text-center text-sm text-muted-foreground">{t("customizeShortcutsRecordingTip")}</div>
-            </div>
-          ) : (
-            <div className="">
-              <div className="flex justify-center mb-4">
-                <Button
-                  variant="outline"
-                  className="font-mono"
-                  onClick={() => {
-                    setIsRecording(true);
-                  }}
-                >
-                  {currentHotkeys[keyName]}
-                </Button>
-              </div>
-              <div className="py-2 text-center text-sm text-muted-foreground">{t("customizeShortcutsTip")}</div>
-            </div>
-          )}
-        </div>
-        <AlertDialogFooter>
-          <Button disabled={!isRecording || !joinedKeys} onClick={changeKeyMap}>
-            {t("save")}
-          </Button>
-          <AlertDialogCancel onClick={clear}>{t("cancel")}</AlertDialogCancel>
-        </AlertDialogFooter>
-      </AlertDialogContent>
-    </AlertDialog>
   );
 };
